+++ packages/e2e/cypress.config.ts
@@ -4,9 +4,9 @@ import { unlinkSync } from 'fs';
 
 export default defineConfig({
     viewportWidth: 1080,
-    defaultCommandTimeout: 5000,
+    defaultCommandTimeout: 10000,
     retries: {
-        runMode: 1,
+        runMode: 2,
         openMode: 0,
     },
     e2e: {